+++ lib/assets/javascripts/cartodb3/editor/layers/basemap-content-view.js
@@ -5,6 +5,7 @@ var BasemapSelectView = require('./basemap-content-views/basemap-select-view');
 var CarouselCollection = require('../../components/custom-carousel/custom-carousel-collection');
 var _ = require('underscore');
 var Backbone = require('backbone');
+var MosaicCollection = require('../../components/mosaic/mosaic-collection');
 
 module.exports = cdb.core.View.extend({
 
@@ -146,13 +147,33 @@ module.exports = cdb.core.View.extend({
       this._selectView.clean();
     }
 
+    this._filteredBasemapsCollection = new MosaicCollection(
+      _.map(this._getFilteredBasemaps(), function (basemap) {
+        return {
+          selected: basemap.get('selected'),
+          val: basemap.get('val'),
+          label: basemap.get('label'),
+          template: basemap.get('template')
+        };
+      }, this)
+    );
+
     this._selectView = new BasemapSelectView({
-      basemapsCollection: this._basemapsCollection,
-      selectedSourceVal: this._getSelectedSource(),
+      filteredBasemapsCollection: this._filteredBasemapsCollection,
       layerDefinitionsCollection: this._layerDefinitionsCollection
     });
     this.addView(this._selectView);
     this.$el.append(this._selectView.render().el);
+  },
+
+  _getFilteredBasemaps: function () {
+    var self = this;
+
+    var filteredBasemaps = this._basemapsCollection.filter(function (mdl) {
+      return mdl.get('category') === self._getCategory();
+    });
+
+    return filteredBasemaps;
   },
 
   _onClickBack: function () {
